Add a `clear` method to `LRUCache` and `StatCache`.
The former way of clearing the cache was to temporarily set its size
to zero (without changing the maximum size, i. e. the `size` attribute
of `LRUCache` objects). The new way is cleaner; the old one was a
workaround which was only tolerated because no one would ever set the
maximum size to zero. ;-) I only do it to implicitly remove all nodes
from the cache in order to make garbage collection more robust.