Reverted to previous version after adding a webserver redirection.
That way, both the local link in the generated `ftputil.html` and
the link on the documentation page will work. If `ftputil.html` is
opened locally in a browser, `ftputil_ru.html` will refer to the file
in the same directory. If the reStructuredText file is copied to the
ftputil website, clicking the link will trigger a redirection to
http://ftputil.sschwarzer.net/trac/wiki/RussianDocumentation.