Don't raise `PermanentError` for `host.path.isdir("/dir/subdir")` if even the directory doesn't exist.
The fix was rather simple, but now ftputil does a lot more behind the
scenes; it traverses up from the path given as argument and does
a stat call on each of these traversals.

This means that in `mock_ftplib.MockSession`, we can no longer have
mixed directory strings for Unix and MS format, so some refactoring of
`MockSession` and changing of some dependent tests was necessary.