In the tests that check if a method returns the same string type it was given, don't test `commonprefix`.
`commonprefix` takes a list of strings, not several string arguments.
What seems to happen is that the underlying Python library code
iterates over the characters of the accidental single string argument.
Although this still does the type comparison, it's misleading in the
test code. `commonprefix` needs its own test method.