Ensure `FTPHost.path.exists` accepts both bytes and unicode strings.
Note that this new requirement is conflicting with the approach
introduced with ticket #53
( http://ftputil.sschwarzer.net/trac/ticket/53 ). That ticket was
about FTP commands failing because of "non-encodable" unicode strings.

The new approach is to allow both bytes and unicode strings as
arguments because this is the behavior of Python 2 and 3 for the
corresponding `os` functions.

Accordingly, all test cases in `test_real_ftp.TestUnicodePaths` will
become obsolete and will eventually be removed. (On the other hand,
new tests are added to make sure that methods take bytes and unicode
strings.)