Use original exception message in `FTPError` (ticket #76).
If we create an `FTPError` instance based on a previously raised
`ftplib` or `socket` exception, take the string representation of this
original exception as the string representation of the new `FTPError`
exception.

The `*args` passed to `FTPError` aren't necessarily enough to
guarantee this. At least for some socket exceptions, `args` is a tuple
of an integer and a string and using only `args[0]` to form the
exception message keeps only the integer.