List contributors in `doc/contributors.txt`.
So far, individual files had copyright notices for contributors.
However, this makes it difficult to properly adapt files in case of
refactoring: If a piece of code is moved to another file, I would need
to find out if this code was contributed by someone else and change
the copyright notice in the target file accordingly.

With the new approach, every file refers to the file
`doc/contributors.txt`, which contains the names of contributors.