Add `FTPHost.listdir` tests for ticket #100
Other methods that take directory or file names have the same issues
as described in the ticket, but at the moment only `listdir` calls
are tested.

There's already a test for non-ASCII file names in `test.test_host`,
but that didn't catch the problem because the mock code in
`test.mock_ftplib` hid the problem.