Change expected error message in unit test
`TestErrorConversion.test_error_message_reuse` tests the construction
of the error message in ftputil. `str(exception)` should include the
error from the caught ftplib exception. In case of the unit test, the
ftplib exception is a socket error (`gaierror`).

It seems the error arguments have changed from `(-2, "Name or service
not known")` to `(-5, "No address associated with hostname")`. I
assume this is due to the update from Fedora 21 to 22. I don't believe
the change is due to a change in Python version because both Python 2
and 3 fail(ed) the unit test and are fixed by the same code change in
`ftputil.error`.