Apply a variant of Roger Demetrescu's patch (see ticket #86)
Thanks, Roger! :-)

The only difference between this changeset and Roger's patch is that
the new code in the changeset has slightly different formatting to
limit the line length to 79 characters.

Contrary to my original intention I haven't added unit tests for the
virtual directory support yet. I'd expect the tests to become rather
convoluted with the current mocking infrastructure in
`test.mock_ftplib`. I'm planning to add tests when more code is
migrated to use the `mock` library (see ticket #98),