Add `__repr__` so that a `StatResult` looks like a named tuple (#91)
Contrary to the suggestion in ticket #91, I can't turn `StatResult`
into a `namedtuple`. Since `StatResult` inherits from `tuple`, its
constructor takes a tuple of the n items to use:

  StatResult((item1, item2, ...))

On the other hand, the constructor of a `namedtuple` takes n arguments
for the tuple items, so the API would become

  StatResult(item1, item2, ...)

Since `StatResult` is a public API for implementation of custom parsers,
I can't just change it.

However, since the main idea behind using a `namedtuple` is its
representation, I added a `__repr__` method so that the representation
of a `StatResult` looks like the representation of a `namedtuple`.

Note that the `__repr__` result is different under Python 2 and 3. The
strings in the `StatResult` are always unicode strings and those have
a `u` prefix under Python 2. I had considered removing the prefix for
Python 2, but I think this could be very confusing during debugging
(including for me).