Explain why we change into the directory to check the existence
In case of virtual directories (#86) we may not see an existining
directory in a listing of the parent directory. Successfully changing
into a directory is a more reliable way to prove its existence.

With the implemented approach, we may still run into problems, if we
can't change into the directory even though it exists. Changing into a
directory `a` may fail if we don't have the permission to do so. In
that case, however, trying to make a child directory `a/b` may fail
anyway.