Fix `TestLstatAndStat.test_repr`
Probably due to a DST change, the expected value for `st_mtime`
changed from 1421705460.0 to 1453241460.0. At the moment, I just
changed the expected value, but in the long run the test should be
changed so that it adapts to a changed DST by itself.

Also changed the expected result for Python 2 to a byte string. I'm
not sure whether it's a good idea to change the string type of `repr`
values "away" from the respective native type.