Fix invalid 226 reply bug (ticket 102)
Under FTP server load, a 226 reply from the server may not be "seen"
while a remote file is closed in `FTPFile.close`. If the reply is only
seen in in the `pwd` call in `FTPHost._available_child`, `ftplib`
raises an `error_reply` exception since no 226 is expected from a
`pwd` call.

This changeset fixes the bug by explicitly catching the `error_reply`
exception. This is the same approach as for timed-out child sessions,
which cause an `error_temp` exception.

The fix will cause `_available_child` to ignore the child session
whose `pwd` caused the `error_reply` exception. However, the child
will be tried again next time `_available_child` is called and should
then have a working `pwd` call (unless the child session timed out,
in which case it will cause an `error_temp`).