Allow alternative exception test in `test_error_message_reuse`
If one of the `socket` functions is called from `ftplib` with a host
string of the form `localhost:21`, this results in a `gaierror`. It
seems the error string for this exception has changed at some point.

Change the unit test so that it works with both the old and the new
exception string.