Fix `test_stat.TestLstatAndStat.test_repr`
The directory entry line that is used for the test is

  -rw-r--r--   1 45854    200          4604 Jan 19 23:11 index.html

Since this line doesn't contain a year, ftputil uses the current year
if January 19th has passed in the current year. Otherwise ftputil will
use the previous year because presumably the date can't be in the
future.

On the other hand, the timestamp `st_mtime` has to use a concrete
datetime value.

Therefore, everytime we get into a new year or time passes to be after
January 19th 23:11, the test needs to be adapted.

A workaround is to use one of the directory lines that have a concrete
year included. However, this way ftputil doesn't see a time anymore,
so this test would not be able to detect whether ftputil took the time
part into account. Then again, the use of the time value is tested
elsewhere.