Apply `normpath` before comparing paths
In both `isfile` and `isdir` of `ftputil.path._Path`, there's a
workaround in case we can't go to a parent directory of the current
directory.

Call `normpath` on the `isfile`/`isdir` path argument before comparing
with `self._host.getcwd()`. This makes sure that we compare correctly
if the paths aren't strictly equal but only semantically equal, i. e.
referring to the same actual file system path.

Don't use `samefile` here because this implicitly tries to call
`stat`. This is exactly what we try to _avoid_ when applying the
workaround.

For now, don't create unit tests because these would be probably
rather awkward.

ticket: 113