Add comment on `set_pasv` call
When I looked at the code, I thought for a moment it was buggy because
`set_pasv` was called even if `use_passive_mode` was `False`. As it
turns out, according to the documentation `set_pasv` can either be
called with `True` (=use passive mode) or `False` (=use active mode).