Handle infinite link chains in `isdir` and `isfile`
If `FTPHost.path.isdir` and `FTPHost.path.isfile` run into an infinite
link chain, return `False`, as the corresponding functions in
`os.path` do.

Before, `isdir` and `isfile` would raise a `PermanentError` with the
information that a recursive link chain was detected. This behavior,
in turn, led to a failure in `FTPHost.walk` (see ticket #107).

ticket: 107