Remove `compat.py`
This module was for Python 2/3 compatibility, similar to the `six`
package. Since the next version of ftputil will only support Python 3,
hardcode the types from `compat.py` in the code that used to use the
`compat` module.

There may still be redundant code that isn't needed when running the
tests under Python 3.