Remove obsolete features
Remove detection of "invalid" features that are either ancient or no
longer invalid. "Ancient" is the use of `xreadlines`; this should
hardly be used anymore nowadays. "No longer invalid" are detections of
byte strings starting with quotes. In Python 3, these are unicode
strings, and it's unlikely that code _deliberately_ uses actual byte
strings starting with `b`.