Disable more Pylint messages
Disable `no-else-returns` for structures like

  if condition:
      return 1
  else:
      return 2

In some cases using `else` here is more "symmetrical" and thus reads
better.

Disable `unnecessary-pass`. This applies mostly to custom exception
classes that have both a docstring and a `pass` after that. I think
the `pass` looks good here, even if Python doesn't require it.

Of course I could disable these messages only for the above cases,
but I think it's not justified to clutter the code with the pragmas
if "violating" the respective rule is harmless.