Add test support class `ScriptedSession`
This code is intended as the replacement of (most of) the currently
used mock code. In the current mock code, multiple tests sometimes
share test data. Hence, a developer must be careful not to add a test
and make other tests fail as a side effect.

The basis of the new mock code is the class
`scripted_session.ScriptedSession`, which is used as (kind of) a
replacement for the session factory argument of `FTPHost`.

Since `ScriptedSession` needs a `script` parameter that common session
factories don't take, I added a helper function
`scripted_session.factory`, which takes a `script` and returns a
session factory that can be used in `FTPHost`.

With the new test code, I could rewrite some of the tests in
`test_host.py` rather nicely. I was able to remove the session class
`UnnormalizedCurrentWorkingDirectory`. However, I had to keep
`FailOnLoginSession`, but only because the class is still used in
another test module.

I'm not yet sure how well the new approach will work with the
remaining tests, but I'm about to find out. I couldn't determine
this just by looking at the test code.