Generalize session factory creation
Allow more than one script in the `factory` function. For each script
in `scripts`, a new `ScriptedSession` with the corresponding `script`
will be used when the factory is called by `FTPHost._make_session`.

The implementation seems to work fine, but it's even more difficult
to understand than the previous version. I'm considering implementing
this with a class.