Shorten `ScriptedSession` tag in output
Instead of printing

  <ScriptedSession at 0x7f45f5aaa780> Expecting method name '__init__' - found it

print a shorter form

  ScriptedSession 1: Expecting method name '__init__' - found it

`ScriptedSession`s get increasing counter values (which is a
class-level variable). This makes the output not only more compact,
but also makes it easier to distinguish different `ScriptedSession`
objects.

Of course, this wouldn't work properly with multithreaded or
multiprocessing code, but I don't see a need for this anytime soon.