Reset session counter for each test
Reset the class variable `_session_count` in `ScriptedSession` every
time a `MultisessionFactory` object is created. In practice, this
means that the counter is reset for each test. However, there could
be a problem when we create more than one `MultisessionFactory` in
a single test.

I added an extra class method `reset_session_count` to hide whether
"resetting" means to set `_session_count` to 0 or to 1.