Use `ScriptedSession` in `test_conditional_upload_with_upload`
This test requires two scripts, one for the host on which
`upload_if_newer` is called and one for the session that is used for
the actual file transfer.

Since `unittest.mock.patch` can't patch the built-in code of
`io.BytesIO` and `io.StringIO`, I added two trivial classes that _can_
be patched.