Revise tests for file reading and writing
In the past, the concepts "ASCII" and "binary" in ftputil referred to
the corresponding concepts for FTP. That is, "binary" meant
transferring the data without changes and "ASCII" meant normalization
of line endings to `\n`.

This meaning of "ASCII" and "binary" is no longer used in ftputil
(since a long time). Instead, we now use the concepts "text" and
"binary" as they're understood in the Python context (i. e. as if
opening files in text or binary mode with `open`).

I also added a few umlauts in some tests to see if the
encoding/decoding works.

That said, I'm sure the tests could be improved:

- Use the same test string in all text/binary read/write tests.

- Test if "half umlauts" are split correctly in binary mode. For
  example, a German "ä" is encoded as b"\xc3\xa4", hence reading
  the first byte should give b"\xc3" and reading the second byte
  should give b"\xa4". On the other hand, if reading one _character_
  from a text file, we'd read the whole single character.