Pass original mode to `socket.makefile`
For the previous Python 2 support, it was necessary to always create a
socket file for binary mode. (There wasn't a text mode for files
created with `socket.makefile`.) Since we no longer support Python 2,
we can pass the mode directly to `socket.makefile` and don't need
wrapping of the file object returned from `socket.makefile` in a
`TextIOWrapper`.

While adapting the tests, it turned out that the file object returned
by `socket.makefile` works slightly different than the one returned
from the previous `TextIOWrapper`. Specifically, `\r\n` sequences are
not converted to `\n` when reading file data.

I'm not exactly sure whether this is a side effect of running the
tests only under Posix. Possibly the behavior is different on Windows.
If the behavior differed depending on the platform, this would be a
bit weird; I wouldn't expect this much less than differences in the
processing of _local_ files (i. e. file system files).
***
Update comment
***
Remove obsolete comment