Add `newline` argument to `StringIO` constructor
The `StringIO` objects are used as the mock result of
`socket.makefile` calls. However, the default for `newline` for
`StringIO` is "\n", which differs from the default `None` for `open`
and `socket.makefile`.

Therefore, to let `StringIO` behave like the real-life `makefile`
return value, we must pass a `newline` argument that matches the
`newline` argument for `host.open` in each of the tests.

If no `newline` argument for `FTPHost.open` is given, the default of
`socket.makefile` is used implicitly, which is `newline=None`, so we
must pass the same value to the `StringIO` constructor explicitly.