Simplify parser test logic
When creating a "dir line" in a `dir_line` call it's (at least for me)
much more intuitive if the time in the resulting string is the local
time (or a naive datetime) I pass to `dir_line`. However, the previous
implementation implicitly applied `time.localtime`, which confused me
a lot during debugging ticket 131.

The changed implementation of `dir_line` takes a naive `datetime`
object and just applies the format, without any timezone changes.

To make everything fit together, change `_test_time_shift` to
internally create the server time as a `datetime` object, not as a
float in seconds.