Order `posixpath` attributes alphabetically
Follow the order in the current Python documentation for `os.path` to
make the list easier to compare with the Python module.

Note that the methods we use from the `posixpath` module can only be
methods that don't need local file system access. Obviously, we need
to access files on the FTP server instead. :-)