Simplify conversion of original exception
In Python 3, exception messages are always unicode strings (`str`),
so special handling for byte string messages is no longer necessary.

When I entered ticket #132, I was under the impression that
`original_exception` was used for exeption chaining and that it could
be removed now. However, `original_exception` is only used to apply
`str(original_exception)` and this use makes sense.