Fix and refactor `parse_unix_time`
While changing `parse_unix_time` to use `datetime` instead of time
tuples, I noticed that time shift wasn't taken into account for the
case where only a year but no time was given.

I noticed that the unix time parser tests had the time shift set to 0,
so these bugs weren't found.

That said, it's still to decide if it makes sense to subtract a time
shift if we now a time only up to day precision. But I think we should
apply the same rules regarding time shift regardless of the presence
of hour and minute. This is just consistent and needs fewer special
rules (the one for the unknown hour:minute case).

TODO: I still need to improve the tests to use a time shift so that
the previous bug is revealed. I want to run these improved tests on
the old code to make sure the bug is present there.

`_mktime` isn't deleted yet because it's also used by the MS line
parser, but I plan to switch this to `datetime` usage, too.