Assume server time without time shift is UTC
By default, assume the timestamps in the FTP server's listings are in
UTC. The user can still set a time shift in case the server doesn't
use UTC timestamps.

The reasons for assuming UTC are explained in ticket #134.

Dealing with "only" UTC and the time shift makes the code a bit easier
to handle because we don't need to consider mismatches between UTC and
local time (let alone different local times on server and client).
That said, we _may_ have time differences between server and client,
but such a time difference should be solely covered by the time shift
value.

Especially tricky is the decision which year to choose for the
timestamp if the directory line from the server doesn't contain a year
value. This can't be solved clearly, so use a heuristics (see the code
comments in `parse_unix_time`).

ticket: 134