Update `set_time_shift` docstring
Emphasize that the client time is UTC. This is consistent with the
new calculation in `_Stat.parse_unix_time`.

In previous versions of ftputil, the time shift was supposed to be the
difference between the local server time (assumed from the timestamp
on the server) and the local client time. However, in some
circumstances the client time was retrieved in UTC (namely by
`os.path.getmtime`), which was inconsistent with other places where
the client time was assumed to be a local time.

ticket: 134