Clear cache if time shift changes
If the time shift changes, all timestamps in the cache will be wrong
with respect to the _new_ time shift. Depending on the cache, reusing
the old timestamps may not be a problem, but in some cases it might
be. Therefore stay on the safe side and clear the cache, sacrificing
efficiency in favor of correct timestamps.

At first sight, we could "correct" the cache entries by applying the
difference between old and new time shift to the existing cache
entries. However, I think the interactions in ftputil are already so
complicated that I'd like to avoid any probably unnecessary
optimizations. Also, the time shift would probably be set soon after
creating the `FTPHost` instance, so anyway only relatively few cache
entries would be affected.