Change `utc_local_time_shift` to `expected_time_shift`
Replace the function `utc_local_time_shift` with
`expected_time_shift`. The purpose of both functions is very similar,
to set the constant `EXPECTED_TIME_SHIFT`. I changed the name because
the time shift is now relative to UTC, not to the local time of the
client. Therefore, the function old name would be confusing.

ticket: 134