Format docstrings and comments
Reformat docstrings and comments to 80 characters. Traditionally, the
margin for the code was at 79 columns, so I chose a bit less for the
comments.

With the switch to Black for formatting the right margin is now at
nearly 90 columns, so it makes sense to also allow a greater right
margin for docstrings and comments. Lines still don't get too long
because the formatted text often starts only in column 5 or 9.