Fix minimum Python version
In order to support path-like objects, we need at least Python 3.6.

It _may_ be possible to create an ftputil version that just doesn't
support path-like objects on Python 3.5, but otherwise works. However,
I think the effort for this wouldn't be justified (especially since
Python 3.5's end-of-life is as close as September 2020, see
https://devguide.python.org/#status-of-python-branches ).