Use base class default if `encoding` isn't given
Make the `encoding` argument of `session_factory` default to `None`.
If `None`, use the default encoding of the base class. This follows
the same logic as for the default handling of the `use_passive_mode`
argument.

Note that the default encoding of `ftplib.FTP` changed from "latin-1"
in Python <= 3.8 to "utf-8" in Python >= 3.9. Therefore, if you want
the same encoding regardless of the Python version, pass your own base
class (compatible with `ftplib.FTP`) which uses a fixed encoding or
pass the `encoding` explicitly to `session_factory`.

ticket: 143