Fix `test_host.py` for Python 3.9
Add an optional argument `encoding` argument to the `as_bytes` helper
function and use this argument in `TestAcceptEitherUnicodeOrBytes`.

The code in `scripted_session` uses the path encoding of the `ftplib`
module in the standard library, whereas the `as_bytes` helper function
in `test_host.py` by default uses the ftputil path encoding default.

Therefore, the tests worked for Python 3.8 and before, but not for
Python 3.9 with the changed path encoding default.

ticket: 143