Allow `encoding` argument only for `FTP` subclasses
If an explicit encoding, i. e. a `str` is passed to `session_factory`,
the `base_class` argument must be `ftplib.FTP` or a subclass of it.

The reason is that we already use a heuristic for Python 3.8 and
earlier vs. Python 3.9 and earlier. This heuristic may work or fail -
possibly silently - for classes that aren't a subclass of
`ftplib.FTP` (including `ftplib.FTP` itself).

With this check in place, the heuristic may look safe, but it actually
isn't if, for example, a class inherits from `ftplib.FTP` in Python
3.9, but overrides the constructor so that it doesn't support the
`encoding` argument anymore.

ticket: 143