Fix regression for custom session factory arguments
The ftputil version 5.0.0 accidentally made it impossible to pass
`args` and `kwargs` from the `FTPHost` constructor to the default
session factory. The problem was that the new default factory was
created with `session.session_factory` and the thus generated
factories don't accept additional arguments.

This change uses `ftplib.FTP` as the default session factory when
running under Python 3.8 and below. (This is the same behavior as in
ftputil versions before 4.0.0.) When running under Python 3.9 and
above, use a `default_session_factory` that inherits from
`ftplib.FTP`, but sets the encoding to
`path_encoding.DEFAULT_ENCODING` if no encoding is specified in the
call of the default session factory.

I decided against extending the classes returned from
`session_factory` so that they would accept arbitrary `args` and
`kwargs`. Such an approach would raise questions like: What should
happen if client code passes a custom `encoding` argument to
`session_factory` and also a (possibly even different) `encoding`
argument to the session factory returned from `session_factory`? I
didn't want to open this can of worms.

Although the actual code change wasn't that difficult, the tests were.

- Testing in `test_real_ftp.py` isn't possible because the test server
  deliberately uses a port above 1024 and the port can't be specified
  in the `ftplib.FTP` constructor.

- The usual mock approach of passing in a mock session factory would
  defeat the purpose of testing the _default_ session factory.

- Mocking parts of the default session factory to use it in `FTPHost`
  tests would be complicated because `FTPHost` uses a lot of the
  session factory API and therefore mocking while at the same time
  satisfying `FTPHost` requirements would be very difficult.

I made the compromise of testing the default session factory itself,
and only for Python 3.9 and above. After all, the default session
factory for Python 3.8 and below is just `ftplib.FTP` and it doesn't
seem worthwhile to adapt the new mock tests so that they also test the
`ftplib.FTP` case.

The current implementation of `default_session_factory` for Python 3.9
and above uses `ftplib.FTP.__init__` instead of `super().__init__`.
The reason is that creating the class derived from `ftplib.FTP` stores
the original class in `default_session_factory.__bases__`, and this
base class isn't affected by mocking `ftplib.FTP`. We would need to
either temporarily modify `__bases__` for the tests or mock `super`.
Maybe I'll try the latter.

ticket: 145