Use `rst2html` "plain" stylesheet
This results in a nicer design than with the default CSS I used so
far.

There's a (probably small) risk that `plain.css` isn't present on the
computer of a user who wants to re-create the documentation, but then
you could say the same for any other software artifacts. In other
words, the user must make sure that `rst2html` is installed, including
`plain.css`.