Raise `FTPIOError` for empty paths ("")
Raising an `IOError` is also the default for Python's native file
system code. Even `touch ""` in the shell reacts with "No such file or
directory".

The exception in the ftputil code is the `FTPHost.chdir` method,
for two reasons:

- `cd ""` in the shell gives no error and just stays in the current
  directory.

- There's one place in the ftputil code where a path is split with
  `head, tail = FTPHost.path.split(path)`. If `path` was just a file
  or directory name, without a preceeding directory, `head` would be
  "" and the following `FTPHost.chdir(head)` would fail.

Since I can imagine this pattern may occur in real-life code and the
shell is "neutral" on it, I'm more defensive here than with the other
ftputil APIs.

ticket: 148