Unwrap `FTPFile` TLS socket before closing it
Without the unwrapping, there may be data "stuck" in the TLS wrapper, so
the underlying socket isn't "flushed" when closing the socket with
`self._conn.close()`.

This problem was raised in ticket #149. Probably it wasn't noticed
earlier, since waiting a short time (one second or even less; where
"waiting" could mean doing anything else that takes a moment) between
the last write to the socket and closing the `FTPFile` seems to transmit
all data. That said, pyftpdlib still reports a broken pipe error in this
case.

I found the solution when checking the current ftplib source code for
`FTP.storbinary`.

With this commit, the workaround of waiting before closing the `FTPFile`
is no longer needed and the broken pipe error in the pyftpdlib log is
gone.

ticket: 149