Always invalidate the cache after attempted removal
When calling `FTPHost.rmdir` or `FTPHost.remove`, always invalidate the
stat cache for the directory or file, respectively.

If an exception occurs, the most likely cause is that the removal
failed, in which case it would be correct to keep the cache entry.
However, if the directory or file was actually deleted on the server,
but something went wrong while finishing the server/client transaction,
keeping the cache entry after the exception would be a mistake.

If we delete the cache entry even if the directory or file wasn't
removed, that shouldn't be a problem because stat queries for the
item will just retrieve the stat information from the server again
(and then we'll know for sure whether the removal succeeded).